Acknowledge exception caused by deleted accounts
Turns out Reddit throws a 404 for deleted accounts. We had to
build in some fucked up weird logic to check for an attribute
present on active accounts. If the attribute (account.name)
isn't present, we return None for the account name so
dependent functions (for now, the submission fetcher)
doesn't get a fright when it tries to grab something that isn't there.

Other minor changes: create an "account_age" attribute for redditor
objects (in the submission fetcher, next commit will include this in
the redditor fetcher), and clean up the .gitignore file.