Reimplement fetching author names sanely.
Turns out one can cast a Redditor object to a string. Who knew?
Either way, this is a cheap and easy way to get a post's author.
This doesn't solve our other problem (author metadata), but resolves
a short-term pain point and gives us more data to work with.

I've also added selftext and domain to the data returned, so we can
do some keyword analytics.