Update index mapping. Fixes #7.
ES needs explicit mapping for fields to be searchable.
Unmapped fields default to fielddata, rendering them unusable for
creating visualizations. We need to map a bunch of fields to the right
data types so we can search efficiently, and leverage X-Pack's graph
search.

I also reverted the structure back to a flat directory, since we're
not quite ready for primetime yet (still running everything in iPython,
lol) and updated the docker-compose config. We now use the Platinum
docker image. X-Pack security is off until I can figure out how to
configure it on the fly.