Add method to stream submissions to Elasticsearch.
This hasn't been tested yet (pending local
Docker stack on Windows) so its existence should be taken with
a pinch of salt for the moment.

We need to maintain a rolling index of posts in the
Elasticsearch index. To do so, we leverage PRAW's
stream() function to run in a loop, grabbing up to
100 posts at a time from a subreddit and indexing
as we go.

Signed-off-by: Dan Urson <dan@plygrnd.net>