Add support for Vault and Consul. Remove env vars for AWS creds.
Managing credentials has, surprisingly, proven to be a giant
pain in the arse throughout this project. We depend on S3 and
Reddit to provide storage and authN respectively. I was storing
Reddit secrets as a JSON blob in S3, which is a) unwieldy as shit,
b) very unsafe should the AWS account be compromised. I've added
Consul (backend storage) and Vault (secret manager) to the
docker-compose configuration. We'll use Vault to store the reddit
creds (username/password/app ID/app secret) and pass these in at
runtime by pulling them from Vault. The tricky part will be
initialising the vault for the first time.