Move back to venv. Disable Vault mlock.
Docker Swarm is missing the ability
to run privileged containers. Fixing
this requires either forking Docker,
or doing weird things to systemd.

Since we're running Tinkerbell in a
single-system context, we disable
mlock so that we can skip using
docker-compose and use the latest,
greatest orchestration mechanism
(Docker Swarm) without messing
about with Compose. This is
also more resilient and better
suited to larger deployments.