Move to Pushshift for historic data fetch.
Reddit turned Cloudsearch off. This sucks
for many reasons, chief among which is that
one can no longer pass timestamps to PRAW's
submissions() method. As such, Tinkerbell
broke, because it relied on that method
to fetch historical data.

We fix this by depending on Pushshift,
a third-party service that indexed the
whole of Reddit and continues to do so.
We can use Pushshift to backfill most
posts, and flip to Reddit's API to stream
new ones into Elasticsearch in real time.

Downside: Pushshift is missing A LOT of
data. This means that we need to backfill
what we have from Reddit at some point.
However, this is very slow. Before we
start backfilling, we need to figure out
a way to parallelize calls to Reddit's API.