V2.0 release with Pushshift integration
We finally have:

- HA Elasticsearch clustering,
- Pushshift integration,
- Working indexing

I had to do some fairly serious fine-tuning
to pick up data from both Pushshift & Reddit.
It turns out that the submission content field
was renamed to "selftext" at some point, which
broke literally everything. I now have it
working, though, which is nice.

We have the basics of a highly-available
Elasticsearch architecture, too. We deploy
an nginx container which fronts the
Elasticsearch cluster and proxies requests to
the master node. This is very, very fragile,
though; at present, adding >1 Elasticsearch
nodes causes Kibana to misbehave, so we need
to fix that. The nginx container is built with
wait-for-it.sh, which will prevent it from
starting until Elasticsearch is ready.

At any rate, it's working now. Mostly. Hooray.