Release v2.0.1 w/HA and scaling
I solved the scaling problem. It was actually
quite simple; I had to disable discovery on
the slave nodes, and ensure that the
Elasticsearch client pointed to the Nginx
container which fronts the master node.

I also tested adding another subreddit to the
compose file (IAmA), which worked well; we now
have two subreddits streaming in!

Finally, I set the default number of ES slave
nodes to 10, so we can ensure that we're
permanently scaled. This also gets around a weird
problem within Docker; when re-deploying a stack,
the scaling per service is reset to 1, which
causes Elasticsearch to shit itself if it was
previously scaled. Fix for this is to manually
scale the slave node service, but that's
suboptimal, and I'd prefer not to put that on
users. So, 10 it is.