Fix monitoring.
I had to enable a proxy_pass option for tcp:9300
so that Kibana could see the slave nodes. It's
hacky and not perfect (nginx will throw a ton of
errors until Kibana and the slaves come up), but
it stabilizes eventually.

TODO: authN/authZ w/SearchGuard, and login
w/Traefik.