Push initial effort to migrate to AWS
This was a huge change. I refactored the "guts" of the project into a single file,
`snoowatch.py` and cleaned up the structure of the classes somewhat so they make sense.
I started to realize that managing the test CFN template by hand would quickly become
difficult, so I started using Troposphere to build it - but this should move to CDK,
since Troposphere is a bit unwieldy.

Additionally, all references to the S3 bucket containing the API creds have been removed.
We're going to push those into Secrets Manager and build a retrieval mechanism for them
instead.

Finally, I removed the docker-compose-secrets-backend.yml and docker-compose.yml config
files. We won't use docker-compose any more; we're moving into Fargate with CFN.