Implement arguments within run_snoowatch
Been noodling on the unusability of the runtime script - it
didn't have the ability to accept the inputs we needed it to expect.
So, I took the easy/lazy way out and used argparse to define the
SM secret/region/subreddit we want.

Of course, this broke everything on which this thing depends,
because the Elasticsearch class is broken as hell. As such,
as of this commit, everything useful is commented out. But...
this is a good starting point.